Split Core. Also Updated all core files to follow PEP8, save a couple of lines that are close to the 80char limit
