+++ dagobah/core/__init__.py
@@ -1,3 +1,7 @@
 from dag import DAG, DAGValidationError
 from .components import JobState, Scheduler, EventHandler
-from .core import Dagobah, Task, Job, DagobahError
+from .dagobah import DagobahError
+from .dagobah import Dagobah
+from .task import Task
+from .jobtask import JobTask
+from .job import Job
