+++ dagobah/core/dagobah_error.py
@@ -1,0 +1,9 @@
+import logging
+
+logger = logging.getLogger('dagobah')
+
+
+class DagobahError(Exception):
+    logger.warn('DagobahError being constructed, something must have gone ' +
+                'wrong')
+    pass
